requirements.txt
selene~=2.0.0rc9
pytest~=8.1.1
allure-pytest~=2.13.5
python-dotenv~=1.0.1
selenium~=4.20.0